renamed kubectl -> kubectl-mtb
